requirements.txt
python-dotenv==1.0.0
pandas==2.2.0
click==8.1.7
pytest==8.0.0
black==24.1.1
flake8==7.0.0
